pulled out api calls from most of the components into a assoiated apihelper file
